PostController Update
Updated the PostController class to use ResponseEntities when returning a value.

From this update, setting up the Frontend should be much easier.

+++ src/main/java/com/morioucho/lifedex/controller/PostController.java
@@ -5,27 +5,37 @@ import com.morioucho.lifedex.service.PostService;
 import com.morioucho.lifedex.service.TrieService;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/posts")
+@RequestMapping("/api/posts")
 public class PostController {
     @Autowired
     private PostService postService;
 
-    @Autowired
-    private TrieService trieService;
-
     @GetMapping
-    public List<Post> getAllPosts(){
-        return postService.findAll();
+    public ResponseEntity<List<Post>> getAllPosts(){
+        List<Post> found = postService.findAll();
+
+        if(found.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+
+        return new ResponseEntity<>(found, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public Post getPostByID(@PathVariable Long id){
-        return postService.getPostByID(id);
+    public ResponseEntity<Post> getPostByID(@PathVariable Long id){
+        Post foundPost = postService.getPostByID(id);
+        if(foundPost == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>(foundPost, HttpStatus.OK);
     }
 
     @PostMapping("/new")
@@ -34,9 +44,15 @@ public class PostController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deletePost(@PathVariable Long id){
-        // TODO: Secure this endpoint so that random posts won't be removed.
+    public ResponseEntity<String> deletePost(@PathVariable Long id){
+        Post foundPost = postService.getPostByID(id);
 
-        postService.deletePost(id);
+        if(foundPost != null){
+            postService.deletePost(id);
+            
+            return new ResponseEntity<>("The post with ID " + id + "was successfully deleted.", HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
